src/ai/rag/requirements.txt
gradio>=4.0.0
databricks-sdk>=0.18.0

